refine chat stream with initial states
