+++ src/lib/scss/custom/components/_alert.scss
@@ -5,6 +5,7 @@
     right: 35%;
     z-index: 8888;
     font-size: 0.9em;
+    white-space: pre-line;
 
     .alert {
         text-align: center;
