+++ src/routes/+layout.svelte
@@ -39,6 +39,7 @@
 		errorUnsubscriber?.();
 	});
 </script>
+
 {#if isLoading}
 	<Loader size={50}/>
 {/if}
